Improved WidgetBackend Manager Service to allow getting all data (doGetFull method) Fixed SystemServicesStatus to be clickable Moved spire deployments fetching to widget backend
